Create random info section
I got most of the funcitonality and look down. Only problem is that I don't
really understand how to change the height of the box. When I change that
margin of the subtext or the image it shrinks the size of the box.
I wonder how I would change the cod eos that the size of the box is consistent.